data-plane/pyproject.toml
[project]
name = "ki2-data-plane"
version = "1.0.0"
description = "KI² Data Plane — unified ingestion, embedding, and search service for municipality RAG pipelines"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    # Framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",

    # Web scraping
    "crawl4ai>=0.8.0",
    "httpx>=0.28.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.3.0",

    # LLM
    "openai>=1.60.0",

    # Storage & caching
    "redis>=5.2.0",
    "qdrant-client>=1.12.0",
    "clickhouse-driver>=0.2.9",

    # Document parsing
    "llama-parse>=0.6.0",
    "pymupdf>=1.24.0",
    "python-docx>=1.1.0",
    "openpyxl>=3.1.0",
    "charset-normalizer>=3.4.0",
    "python-multipart>=0.0.17",

    # File discovery
    "smbprotocol>=1.14.0",
    "ldap3>=2.9.0",
    "boto3>=1.35.0",

    # Observability
    "prometheus-client>=0.21.1",
    "structlog>=24.4.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-httpx>=0.35.0",
]
